Updated a bunch of functions to use SpellID enum instead of uin16

+++ src/map/entities/automatonentity.cpp
@@ -180,7 +180,7 @@ void CAutomatonEntity::OnCastFinished(CMagicState& state, action_t& action)
     auto PSpell = state.GetSpell();
     auto PTarget = static_cast<CBattleEntity*>(state.GetTarget());
 
-    PRecastContainer->Add(RECAST_MAGIC, PSpell->getID(), action.recast);
+    PRecastContainer->Add(RECAST_MAGIC, static_cast<uint16>(PSpell->getID()), action.recast);
 
     if (PSpell->tookEffect())
     {
